spring-healthcare-data/src/test/java/ahrnuld/springhealthcare/services/map/OwnerServiceMapTest.java
package ahrnuld.springhealthcare.services.map;

import ahrnuld.springhealthcare.model.Owner;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.Set;

import static org.junit.jupiter.api.Assertions.*;

class OwnerServiceMapTest {

    OwnerServiceMap ownerServiceMap;
    final Long ownerId = 1l;
    final String lastName = "Smith";

    @BeforeEach
    void setUp() {
        ownerServiceMap = new OwnerServiceMap(new PetTypeServiceMap(), new PetServiceMap());
        Owner anOwner = new Owner();
        anOwner.setId(1L);
        anOwner.setLastName(lastName);
        ownerServiceMap.save(anOwner);
    }

    @Test
    void findAll() {
        Set<Owner> ownerSet = ownerServiceMap.findAll();
        assertEquals(1, ownerSet.size());
    }

    @Test
    void findById() {
        Owner owner = ownerServiceMap.findById(ownerId);
        assertEquals(ownerId, owner.getId());
    }

    @Test
    void save() {
        Owner owner = new Owner();
        owner.setId(ownerId);
        owner = ownerServiceMap.save(owner);

        assertEquals(ownerId, owner.getId());
    }

    @Test
    void saveNoId() {
        Owner owner = new Owner();
        owner = ownerServiceMap.save(owner);
        assertNotNull(owner.getId());

    }

    @Test
    void delete() {
        ownerServiceMap.delete(ownerServiceMap.findById(ownerId));
        assertEquals(0, ownerServiceMap.findAll().size());
    }

    @Test
    void deleteById() {
        ownerServiceMap.deleteById(ownerId);
        assertEquals(0, ownerServiceMap.findAll().size());
    }

    @Test
    void findByLastName() {
        Owner owner = ownerServiceMap.findByLastName(lastName);
        assertNotNull(owner);
    }
}